// src/utils/open-im-sdk-wasm/types/entity.d.ts
export declare type CardElem = {
    userID: string;
    nickname: string;
    faceURL: string;
    ex: string;
};