// src/utils/open-im-sdk-wasm/types/params.d.ts
export declare type CustomMsgParams = {
    data: string;
    extension: string;
    description: string;
};
// ...
export declare type MergerMsgParams = {
    messageList: MessageItem[];
    title: string;
    summaryList: string[];
};